tweetrepository conflict

+++ src/main/java/com/example/bitter/repository/TweetRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TweetRepository extends JpaRepository<Tweet, Integer> {
+public interface TweetRepository extends JpaRepository<Tweet, Long> {
     List<Tweet> findByDeletedFalseOrderByPosted();
 }
